Add negative pixel count handling to support reversed layouts
Our Fade Candy setup has two branching LED strips which start at
the middle, and we'd like to map them to a single channel starting
at one end and going all the way to the other.

To reverse the order of one of the strips without this change we'd
need to add individual map entries for all of the pixels in that
strip. Instead, we want to reverse the order of the pixels in that
segment.