Tidy up some variables and update doc/comments
Add references to the +/- pixel count option to the comment in code and update the markdown doc for the configuration file format to describe how to use it.

The channel variable was also getting shadowed inside of opcMapPixelColors in the color swizzling loop, so I renamed the configuration variable to mapChannel and marked it as const.